pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Second python project on Hexlet"
authors = ["Ilya Dutlov <id@dutlov.ru>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = 'gendiff.scripts.main:main'